update production env to use local paths
our production single-host setup uses an nginx proxy to direct traffic
to the appropriate servies. This means that the frontend does not need
to have a separate url to talk to the graphql or blob servers, and can
just put in a path.  Since this file is not giving any secrets, and is
already committed, we can change it to better fit the new setup and have
it be used in our docker image build.